+++ internal/storage/vscode-tutorial.code-workspace
@@ -1,0 +1,17 @@
+{
+	"folders": [
+		{
+			"path": "../../../../vscode-tutorial"
+		},
+		{
+			"path": "../../../../url-shortener"
+		},
+		{
+			"path": "../../../../pet-code"
+		},
+		{
+			"path": "../../.."
+		}
+	],
+	"settings": {}
+}
